Basic Enemy Spawner
Forgot to push this the other day. Made an enemy spawner that will spawn enemies at a regular interval. Made enemies that follow the host player (will update to marked player when that mechanic is implemented). Enemies die to bullets. Forcefield does nothing for now, as the attack seems unfinished.