NavMesh implemented; updated enemy targeting and damaging
Created a variant of level 1 to receive navmesh and added spawns where I felt appropriate (feel free to change them). Created a navmesh object to be placed in the scene alongside the level object. Simple press bake once both are in a scene to make the navmesh. 

Enemies will now use navmesh for navigation and will target the marked player. If no player is marked, they will simply stay still and wait for a player to be marked. They now also deal and take damage when colliding with players. 

A minor conflict occurred between the last push and my files, but was easy to resolve. Enemies check for death how they did prior to this update to avoid unnecessary operations during Update(), and a small helper function named Die() was created to be overridden by the exploding enemy.